+++ aicir/qas/CRLQAS.py
@@ -23,8 +23,8 @@ import numpy as np
 import torch
 import torch.nn as nn
 
-from ..channel.backends.numpy_backend import NumpyBackend
-from ..channel.operators import Hamiltonian
+from ..backends.numpy_backend import NumpyBackend
+from ..operators import Hamiltonian
 from ..core.circuit import Circuit
 from ..core.gates import gate_to_matrix
 from ..core.state import State